+++ banano-quest/banano-quest/Models/Quest+CoreDataProperties.swift
@@ -10,6 +10,7 @@
 import Foundation
 import CoreData
 
+public typealias QuestListCompletionHandler = (_: [Quest]?, _: Error?) -> Void
 
 extension Quest {
 
@@ -26,4 +27,24 @@ extension Quest {
     @NSManaged public var winners: Winners?
     @NSManaged public var metadata: Metadata?
 
+}
+
+extension Quest {
+    public static func retrieveQuestList(handler: @escaping QuestListCompletionHandler) throws {
+        // Quests list retrieved from CoreData
+        var quests = [Quest]()
+        
+        let fetchRequest = NSFetchRequest<Quest>(entityName: "Quest")
+        // Sync quest list
+        Networking.getQuestList { (error) in
+            do {
+                // Retrieve quest list from coreData
+                quests = try BaseUtil.mainContext.fetch(fetchRequest) as [Quest]
+                handler(quests,nil)
+            }
+            catch let error as NSError {
+                handler(nil,error)
+            }
+        }
+    }
 }
