+++ banano-quest/banano-quest/Utils/BaseUtil.swift
@@ -1,0 +1,22 @@
+//
+//  BaseUtil.swift
+//  banano-quest
+//
+//  Created by Pabel Nunez Landestoy on 7/4/18.
+//  Copyright © 2018 Michael O'Rourke. All rights reserved.
+//
+
+import Foundation
+import CoreData
+import UIKit
+
+public class BaseUtil {
+    public static var mainContext: NSManagedObjectContext {
+        get {
+            guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else {
+                return NSManagedObjectContext.init(concurrencyType: .mainQueueConcurrencyType)
+            }
+            return appDelegate.persistentContainer.viewContext
+        }
+    }
+}
