+++ banano-quest/banano-quest/Utils/Networking.swift
@@ -10,7 +10,7 @@ import Foundation
 import PocketEth
 import Pocket
 
-public typealias QuestListHandler = (_: [Quest]?, _: Error?) -> Void
+public typealias QuestListHandler = (_: Error?) -> Void
 public typealias NewQuestHandler = (_: Quest?, _: Error?) -> Void
 public typealias QuestCompletionHandler = (_: QueryResponse?, _: Error?) -> Void
 public typealias QuestCountHandler = (_: Int32?, _: Error?) -> Void
@@ -18,36 +18,40 @@ public typealias QuestCountHandler = (_: Int32?, _: Error?) -> Void
 public class Networking {
     static let nodeURL = URL.init(string: "https://node.url")
     
-    public static func getQuestList() {
+    public static func getQuestList(handler: @escaping QuestListHandler) {
         var params = [AnyHashable : Any]()
-        
+        // Retrieve Quests count
         getQuestListCount(handler: { (count, error) in
             if error != nil {
                 print("\(String(describing: error))")
             }
             
             var index = count ?? 0 - 1
-            
+            // Create and execute a Query for each Quest from most recent to older
             while index >= 0 {
                 params["_tokenAddress"] = "tokenaddress"
                 params["_questIndex"] = index
                 
                 do {
                     let query = try createQuery(with: params)
-                    
+                    // Query execution to retrieve a single quest
                     Pocket.getInstance(pocketNodeURL: nodeURL!).executeQuery(query: query, handler: { (response, error) in
                         if error != nil {
                             print("Failed to get Pocket instance for getQuestList() with error:\(String(describing: error))")
                         }else{
-                            _ = QuestDownloadParsing.parseDownload(dict: response!)
+                            // Quest parsing
+                            let quest = QuestDownloadParsing.parseDownload(dict: response!)
+                            // Quest is saved into CoreData
+                            quest.save()
                         }
                     })
                 } catch let error as NSError{
-                    print(error)
-                    return
+                    handler(error)
                 }
                 index = index - 1
             }
+            // Completion handler
+            handler(nil)
         })
     }
     
@@ -59,6 +63,7 @@ public class Networking {
         params["_tokenAddress"] = "tokenaddress"
         
         do {
+            // Create and execute a Query to retrieve quest list count
             let query = try createQuery(with: params)
             
             Pocket.getInstance(pocketNodeURL: nodeURL!).executeQuery(query: query, handler: { (response, error) in
